requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart==0.0.6
pydantic>=2.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=10.0.0
pandas>=2.0.0
openpyxl>=3.1.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv>=1.0.0
google-cloud-vision>=3.4.4
pdf2image>=1.16.3
poppler-utils==0.1.0  # Required for pdf2image 
numpy>=1.24.0 